remove repeated debug logs when ostentus missing
This commit removes the debug message sent each time a libostentus
function is called but the Ostentus board is not present. These messages
clutter the logs and are not a graceful way to fail. There is still an
error message during initialization when Ostentus is not present. The
user can also view the return code which will be `-EFAULT` when calling
libostentus functions without an Ostentus connected.

Signed-off-by: Mike Szczys <szczys@hotmail.com>